Add note clarifying that Greasemonkey is not supported
When JumpFlowy is loaded as a user script in Greasemonkey, the user
script appears to not able to see global variables which are added
after the script is loaded. Given that Workflowy sometimes adds
important global variables *after* the user script is loaded
(e.g. 'project_tree'), it's not clear how to reliably access these
variables from within a Greasemonkey user script.