Improve navigation to nodes
In openNodeHere function, use the location object to navigate to a node
rather than creating and opening a URL.

Add isWorkFlowyHomeUrl function.

When following a node that has a URL to the root node, use the
location object to navigate to it, to prevent a page reloads and
unwanted extra items in the location history.

Don't include a node's note text in nodeToSearchTermText function.