Speed up bookmarks search
Speed up doesItemHaveTag method by eliminating the most common case first
(items which have no tags at all), by checking for the presence of the
first character of the tag (# or @) before converting to plain text.

Note: this speed up based on the assumption that a large number of items
have no tags.