Reformat code to max line length 120 (up from 80) for readability
No functional changes.

This code base intentionally contains descriptive (long) variable and
function names, opting for explicitness (at the cost of verbosity) over
terse names which hide detail. For that reason, the 80 char limit breaks
up too many lines which read better as a single line.

A comparison was done between the 80, 100, and 120 max options for this
code, and the 120 limit yielded the most readable code.